streamlit_app/requirements.txt
streamlit>=1.35.0
python-dotenv>=1.0.0
httpx>=0.27.0
pandas>=2.0.0
